imagen_a_texto/requirements.txt
pytesseract
Pillow
python-docx
tkinter
